update (mainly fixes and change libbase type to struct ReqToolsBase *ReqToolsBase)
git-svn-id: https://svn.aros.org/svn/aros/trunk@9213 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/reqtools/misc.c
@@ -141,7 +141,7 @@ ULONG LoopReqHandler(struct rtHandlerInfo *rthi)
 	else
 	    sigs = Wait(rthi->WaitMask);
 
-	handler_retval = rtReqHandlerA(rthi, sigs, NULL);
+	handler_retval = (ULONG)rtReqHandlerA(rthi, sigs, NULL);
 	
     } while (handler_retval == CALL_HANDLER);
